--- agents/plwr_test.py
@@ -1,49 +1,0 @@
-code_to_execute = '''
-from playwright.async_api import async_playwright
-from playwright._impl._errors import TimeoutError
-
-async def main():
-    try:
-        # Start Playwright in async mode
-        async with async_playwright() as p:
-            # Launch the browser
-            browser = await p.chromium.launch(headless=False)
-            page = await browser.new_page()
-
-            # Login as campaign user
-            await page.goto('http://localhost:5173/login')
-            await page.fill('input[type="email"]', 'frontend.feedback@campaign')
-            await page.fill('input[type="password"]', '123')
-            await page.click('button[type="submit"]')
-            await page.wait_for_url('**/')
-            #await page.wait_for_load_state('networkidle')
-
-            # Navigate to campaign profile page
-            await page.goto('http://localhost:5173/campaign/test-campaign-uuid')
-
-            # Ensure all dynamic content is loaded
-            await page.wait_for_load_state('networkidle')
-
-            # Take a screenshot
-            output = await page.screenshot()
-
-            # Close the browser
-            await browser.close()
-
-    except Exception as e:
-        output = f"{type(e).__name__}: {e}"
-
-    return output
-
-# Run the async function
-import asyncio
-result = asyncio.run(main())
-print(result)
-
-
-
-'''
-
-# Execute the code using exec()
-exec(code_to_execute)
-exec(code_to_execute)
