+++ manual_tests/executor_scenario0.py
@@ -1,4 +1,4 @@
-from agents.executor_agent import Executor
+from src.agents.executor_agent import Executor
 import shutil
