+++ manual_tests/executor_scenario1.py
@@ -3,7 +3,7 @@ current = os.path.dirname(os.path.realpath(__file__))
 parent = os.path.dirname(current)
 sys.path.append(parent)
 from manual_tests.utils_for_tests import setup_work_dir, cleanup_work_dir
-from agents.executor_agent import Executor
+from src.agents.executor_agent import Executor
 from dotenv import load_dotenv, find_dotenv
