+++ single_task_coder.py
@@ -1,21 +1,21 @@
 if __name__ == "__main__":
-    from utilities.graphics import print_ascii_logo
+    from src.utilities.graphics import print_ascii_logo
     print_ascii_logo()
 from dotenv import find_dotenv
-from utilities.set_up_dotenv import set_up_env_coder_pipeline
+from src.utilities.set_up_dotenv import set_up_env_coder_pipeline
 if not find_dotenv():
     set_up_env_coder_pipeline()
 
-from agents.researcher_agent import Researcher
-from agents.planner_agent import planning
-from agents.executor_agent import Executor
-from agents.debugger_agent import Debugger
-from agents.frontend_feedback import write_screenshot_codes, execute_screenshot_codes
+from src.agents.researcher_agent import Researcher
+from src.agents.planner_agent import planning
+from src.agents.executor_agent import Executor
+from src.agents.debugger_agent import Debugger
+from src.agents.frontend_feedback import write_screenshot_codes, execute_screenshot_codes
 import os
-from utilities.user_input import user_input
-from utilities.print_formatters import print_formatted
-from utilities.start_project_functions import set_up_dot_clean_coder_dir
-from utilities.util_functions import create_frontend_feedback_story
+from src.utilities.user_input import user_input
+from src.utilities.print_formatters import print_formatted
+from src.utilities.start_project_functions import set_up_dot_clean_coder_dir
+from src.utilities.util_functions import create_frontend_feedback_story
 from concurrent.futures import ThreadPoolExecutor
