+++ src/prompts/manager_system.prompt
