+++ src/prompts/researcher_file_answerer.prompt
