+++ src/prompts/researcher_system.prompt
