+++ src/prompts/researcher_tc_system.prompt
