+++ src/prompts/voter_system.prompt
