+++ src/utilities/manager_utils.py
@@ -1,7 +1,7 @@
 from langchain_openai.chat_models import ChatOpenAI
 from langchain_community.chat_models import ChatOllama
 from langchain_anthropic import ChatAnthropic
-from utilities.llms import llm_open_router
+from src.utilities.llms import llm_open_router
 from langchain_core.output_parsers import StrOutputParser
 from langchain_core.prompts import PromptTemplate
 from todoist_api_python.api import TodoistAPI
